created crates `sp_protos` and `sp_fragments` + re-factored all Cargo.toml files
