+++ pallets/aliases/Cargo.toml
@@ -10,116 +10,33 @@ version = '0.0.1'
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
 
-[dependencies.codec]
-default-features = false
-features = ['derive']
-package = 'parity-scale-codec'
-version = '3.0.0'
-
-[dependencies.frame-benchmarking]
-default-features = false
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '4.0.0-dev'
-optional = true
-
-[dependencies.frame-support]
-default-features = false
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '4.0.0-dev'
-
-[dependencies.frame-system]
-default-features = false
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '4.0.0-dev'
-
-[dependencies.scale-info]
-default-features = false
-features = ['derive']
-version = '2.0'
-
-[dependencies.sp-core]
-default-features = false
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '6.0.0'
-
-[dependencies.sp-io]
-default-features = false
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '6.0.0'
-
-[dependencies.sp-std]
-default-features = false
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '4.0.0-dev'
-
-[dependencies.sp-runtime]
-default-features = false
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '6.0.0'
-
-[dependencies.pallet-protos]
-default-features = false
-path = '../protos'
-
-[dependencies.pallet-fragments]
-default-features = false
-path = '../fragments'
-
-[dependencies.pallet-detach]
-default-features = false
-path = '../detach'
-
-[dependencies.pallet-assets]
-default-features = false
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '4.0.0-dev'
-
-[dependencies.pallet-balances]
-default-features = false
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '4.0.0-dev'
-
-[dependencies.pallet-accounts]
-default-features = false
-path = '../accounts'
-
-[dependencies.pallet-proxy]
-default-features = false
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '4.0.0-dev'
-
-[dependencies.pallet-timestamp]
-default-features = false
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '4.0.0-dev'
-
-[dependencies.pallet-contracts]
-default-features = false
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '4.0.0-dev'
-
-[dependencies.pallet-oracle]
-default-features = false
-path = '../oracle'
-
-[dependencies.pallet-clusters]
-default-features = false
-path = '../clusters'
-
 [dependencies]
+sp-core = { version = '6.0.0', default-features = false, git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
+sp-io = { version = '6.0.0', default-features = false, git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
+sp-std = { version = '4.0.0-dev', default-features = false, git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
+sp-runtime = { version = '6.0.0', default-features = false, git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
 sp-clamor = { version = '0.1.0', path = '../../primitives/clamor', default-features = false }
+sp-fragments = { version = '0.1.0', path = '../../primitives/fragments', default-features = false }
+
+frame-benchmarking = { version = '4.0.0-dev', default-features = false, optional = true, git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
+frame-support = { version = '4.0.0-dev', default-features = false, git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
+frame-system = { version = '4.0.0-dev', default-features = false, git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
+pallet-randomness-collective-flip = { version = '4.0.0-dev', default-features = false, git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
+pallet-balances = { version = '4.0.0-dev', default-features = false, git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
+pallet-assets = { version = '4.0.0-dev', default-features = false, git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
+pallet-proxy = { version = '4.0.0-dev', default-features = false, git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
+pallet-timestamp = { version = '4.0.0-dev', default-features = false, git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
+pallet-contracts = { version = '4.0.0-dev', default-features = false, git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
+pallet-protos = { default-features = false, path = '../protos' }
+pallet-fragments = { default-features = false, path = '../fragments' }
+pallet-accounts = { default-features = false, path = '../accounts' }
+pallet-detach = { default-features = false, path = '../detach' }
+pallet-oracle = { default-features = false, path = '../oracle' }
+pallet-clusters = { default-features = false, path = '../clusters' }
+
+codec = { package = 'parity-scale-codec', version = '3.0.0', default-features = false, features = ['derive'] }
+scale-info = { version = '2.0', default-features = false, features = ['derive'] }
+
 hex = { version = "0.4.3", default-features = false }
 serde = { version = "1.0.136", default-features = false, features = ["derive"], optional = true }
 serde_json = { version = '1.0.79', default-features = false, features = ['alloc'] }
@@ -127,32 +44,26 @@ protos = { version = "0.1.21", default-features = false }
 log = { version = "0.4.14", default-features = false }
 bitflags = "1.3.2"
 
-[dependencies.pallet-randomness-collective-flip]
-default-features = false
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '4.0.0-dev'
-
 [features]
 default = ["std"]
 std = [
-	'serde',
-	"codec/std",
-	"scale-info/std",
-	"pallet-protos/std",
-	"pallet-fragments/std",
-	"pallet-clusters/std",
-	"frame-support/std",
-	"frame-system/std",
-	"frame-benchmarking/std",
+	"sp-core/std",
 	"sp-io/std",
 	"sp-std/std",
 	"sp-runtime/std",
-	"sp-core/std",
+	"frame-benchmarking/std",
+	"frame-support/std",
+	"frame-system/std",
 	"pallet-randomness-collective-flip/std",
 	'pallet-balances/std',
 	'pallet-assets/std',
 	'pallet-contracts/std',
+	"pallet-protos/std",
+	"pallet-fragments/std",
+	"pallet-clusters/std",
+	"codec/std",
+	"scale-info/std",
+	'serde',
 	'protos/std'
 ]
 runtime-benchmarks = ["frame-benchmarking/runtime-benchmarks"]
