+++ pallets/detach/src/lib.rs
@@ -80,7 +80,8 @@ use sp_io::{crypto as Crypto, hashing::keccak_256, offchain_index};
 use sp_runtime::{offchain::storage::StorageValueRef, traits::Hash, MultiSigner};
 use sp_std::{collections::btree_set::BTreeSet, vec, vec::Vec};
 
-use sp_clamor::{Hash128, Hash256, InstanceUnit};
+use sp_clamor::{Hash128, Hash256};
+use sp_fragments::{InstanceUnit};
 
 use frame_system::offchain::{
 	AppCrypto, CreateSignedTransaction, SendUnsignedTransaction, SignedPayload, Signer,
