+++ pallets/fragments/rpc/Cargo.toml
@@ -10,42 +10,22 @@ version = '0.0.1'
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
+sp-runtime = { version = '6.0.0', git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
+sp-clamor = { version = '0.1.0', path = '../../../primitives/clamor', default-features = false }
+sp-api = { version = '4.0.0-dev', git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
+sp-rpc = { version = '6.0.0', git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
+sp-blockchain = { version = '4.0.0-dev', git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
+
+pallet-fragments = { version = '0.0.1', default-features = false, path = '..' }
+
 codec = { package = "parity-scale-codec", version = "3" }
+
+pallet-fragments-rpc-runtime-api = { version = '0.0.1', path = '../rpc/runtime-api' }
+
 jsonrpsee = { version = "0.15.1", features = ["server", "macros"] }
 serde = { version = "1", features = ["derive"] }
-sp-clamor = { version = '0.1.0', path = '../../../primitives/clamor', default-features = false }
 hex = "^0.4.3"
 protos = { version = "0.1.21", default-features = false }
 
-# Substrate Dependencies
-[dependencies.sp-api]
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '4.0.0-dev'
-
-[dependencies.sp-runtime]
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '6.0.0'
-
-[dependencies.sp-rpc]
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '6.0.0'
-
-[dependencies.sp-blockchain]
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '4.0.0-dev'
-
-[dependencies.pallet-fragments-rpc-runtime-api]
-path = '../rpc/runtime-api'
-version = '0.0.1'
-
 [dev-dependencies]
 serde_json = "1"
-
-[dependencies.pallet-fragments]
-default-features = false
-path = '..'
-version = '0.0.1'
