+++ pallets/protos/rpc/runtime-api/Cargo.toml
@@ -11,40 +11,22 @@ version = '0.0.1'
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
+sp-std = { version = '4.0.0-dev', default-features = false, git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
+sp-runtime = { version = '6.0.0', default-features = false, git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
+sp-clamor = { version = '0.1.0', path = '../../../../primitives/clamor', default-features = false }
+sp-api = { version = '4.0.0-dev', default-features = false, git = 'https://github.com/fragcolor-xyz/substrate.git', tag = 'fragnova-v0.0.6' }
+
+pallet-protos = { version = '0.0.1', default-features = false, path = '../..' }
+
 codec = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
 scale-info = { version = "2.0", default-features = false, features = ["derive"] }
-sp-clamor = { version = '0.1.0', path = '../../../../primitives/clamor', default-features = false }
-
-# Substrate Dependencies
-[dependencies.sp-api]
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '4.0.0-dev'
-default-features = false
-
-[dependencies.sp-runtime]
-default-features = false
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '6.0.0'
-
-[dependencies.sp-std]
-default-features = false
-git = 'https://github.com/fragcolor-xyz/substrate.git'
-tag = 'fragnova-v0.0.6'
-version = '4.0.0-dev'
-
-[dependencies.pallet-protos]
-default-features = false
-path = '../..'
-version = '0.0.1'
 
 [features]
 default = ["std"]
 std = [
+	"sp-std/std",
+	"sp-runtime/std",
 	"sp-api/std",
 	"codec/std",
 	"scale-info/std",
-	"sp-std/std",
-	"sp-runtime/std",
 ]
